+++ CHANGELOG.md
@@ -12,9 +12,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
+ - Added extension methods:
+   - TypeExtensions:
+     - `Type.Is<T>()`: Reflection alternative to `someObj is SomeType`. Usage would be `someObj.GetType().Is<SomeType>()`.
 ### Changed
 ### Removed
 ### Fixed
+
+
+
+## [v1.1.2] - 2024-04-24
+### Added
+ - Added extension methods:
+   - TypeExtensions:
+     - `Type.Is<T>()`: Reflection alternative to `someObj is SomeType`. Usage would be `someObj.GetType().Is<SomeType>()`.
 
 
 
@@ -104,7 +115,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 
-[unreleased]: https://github.com/ash-dvlpr/cinder-utils/compare/v1.1.1...master
+[unreleased]: https://github.com/ash-dvlpr/cinder-utils/compare/v1.1.2...master
+[v1.1.2]: https://github.com/ash-dvlpr/cinder-utils/releases/tag/v1.1.2
 [v1.1.1]: https://github.com/ash-dvlpr/cinder-utils/releases/tag/v1.1.1
 [v1.1.0]: https://github.com/ash-dvlpr/cinder-utils/releases/tag/v1.1.0
 [v1.0.2]: https://github.com/ash-dvlpr/cinder-utils/releases/tag/v1.0.2